% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kSearch.R
\name{summary.kSearch}
\alias{summary.kSearch}
\title{Summarizing an Object of Class kSearch}
\usage{
\method{summary}{kSearch}(object)
}
\arguments{
\item{object}{An object of class \code{kSearch}.}
}
\description{
Summarizes a \code{kSearch} object.
}
\examples{
n <- 500
S <- cbind(rnorm(n, sd = 2), rnorm(n, sd = 1.5), rnorm(n), rnorm(n))
A <- rorth(4)
X <- S \%*\% t(A)
result <- kSearch(X, method = PCAasymp)
summary(result)

}
\seealso{
\code{\link{print.kSearch}}, \code{\link{kSearch}}
}
\author{
Katariina Perkonoja
}
\keyword{methods}
\keyword{print}
