\name{ggscreeplot}
\alias{ggscreeplot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
ggplot2-style screeplot
}
\description{
A generic method for ggplot2-style screeplots.
}
\usage{
ggscreeplot(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{An object of an appropriate class.}
  \item{...}{Additional arguments.}
}


\author{
Joni Virta, Klaus Nordhausen
}

\keyword{ hplot }
